Move integer parse check to UI.
