+++ test/GameTests.java
@@ -38,21 +38,6 @@ public class GameTests {
     }
 
     @Test
-    public void checkSelectionIsValid() {
-        assertEquals(true, game.validSelection("1"));
-    }
-
-    @Test
-    public void checksSelectionIsNotString() {
-        assertEquals(false, game.validSelection("one"));
-    }
-
-    @Test
-    public void checksSelectionIsValidInteger() {
-        assertEquals(false, game.validSelection("9999999999999999999999999999999999999"));
-    }
-
-    @Test
     public void checksSelectionIsNotOffTheBoard() {
         assertEquals(false, game.selectionOnBoard(10));
     }
